requirements.txt
pyrogram~=1.2.9
TgCrypto
Pillow~=8.3.1
requests~=2.25.1
SQLAlchemy~=1.4.21
psycopg2